feture: divide tests
