+++ src/services/index.ts
@@ -1,6 +1,6 @@
 import { IUser } from '../interfaces/index';
 import { decodeToken } from '../utils';
-import { register } from './userService';
+import { confirm, getUser, register } from './userService';
 
 export const registerUser = async (payload: IUser) => {
   if (!payload) {
@@ -10,6 +10,8 @@ export const registerUser = async (payload: IUser) => {
 };
 
 export const confirmUserEmail = async (payload: any) => {
-  const email = decodeToken(payload.token);
- 
+  const data = decodeToken(payload.token);
+  await confirm(data.email);
+  const userData = await getUser(data.email);
+  return userData;
 };
